refactor: split read_rng from read_simulator
