+++ mjolnir/input/read_simulator.cpp
@@ -6,10 +6,6 @@
 
 namespace mjolnir
 {
-template RandomNumberGenerator<SimulatorTraits<double, UnlimitedBoundary>       > read_rng(const toml::value&);
-template RandomNumberGenerator<SimulatorTraits<float,  UnlimitedBoundary>       > read_rng(const toml::value&);
-template RandomNumberGenerator<SimulatorTraits<double, CuboidalPeriodicBoundary>> read_rng(const toml::value&);
-template RandomNumberGenerator<SimulatorTraits<float,  CuboidalPeriodicBoundary>> read_rng(const toml::value&);
 
 // ----------------------------------------------------------------------------
 // read_molecular_dynamics_simulator
